Initial Page File Handling Added
Sample files added to page, only a static pag WebResponse class has been created, WebResponses can be extended to true server side processes, new Web Responses need to be pushed into the pages array to be added to the page. Might not be the best way if the page gets larger as currently each request requires a linear search.